msm: kgsl: 'qcom,kgsl-3d0' device does not need to get regulators
The vddcx-supply and vdd-supply properties are not defined by the
'qcom,kgsl-3d0' device in GPU DTSI since the corresponding regulators
are not required by the device. Invoking devm_regulator_get for vdd and
vddcx supply results in obtaining a dummy supply. Only invoke
devm_regulator_get for 'qcom,kgsl-3d0' device when the corresponding
supply properties are defined in the DTSI.

Change-Id: Ie3ae363e5644ff123bc984359abef9af2dc102d9
Signed-off-by: Raghu Ananya Arabolu <rarabolu@codeaurora.org>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>